+++ src/main/java/com/endava/calculator/expert/ExpertOperations.java
@@ -6,7 +6,7 @@ public interface ExpertOperations extends BasicOperations {
 
     abstract double pow(int base, int exponent);
     abstract double root(int a);
-    abstract long fact(int n);
+    abstract long fact(double n);
     abstract long factRec(int n);
     abstract double calculate (String s);
